requirements.txt
python-multipart
python-dotenv
fastapi
uvicorn
openai
anthropic
filetype

# demo
streamlit